rename pkg
