+++ agent/components/vectorstores/faiss_builder.py
@@ -1,7 +1,7 @@
 import faiss
 from langchain_community.vectorstores import FAISS
-from agent.server.embeddings import OpenAIEmbBuilder
-from agent.server.knowledges import load_all_pdf_knowledges
+from agent.components.embeddings import OpenAIEmbBuilder
+from agent.components.knowledges import load_all_pdf_knowledges
 
 class FaissBuilder():
     def __init__(self, embeddings, documents = None, local_index_path=None):
