+++ pyproject.toml
@@ -14,6 +14,7 @@ langchain-community = "^0.3.1"
 langchain-openai = "^0.2.1"
 faiss-cpu = "^1.8.0.post1"
 pypdf = "^5.0.0"
+streamlit = "^1.38.0"
 
 
 [build-system]
